added ashell configs, bashrc, edits
added:
- ashell configs:
a. the current config - config.toml
b. the default config - default.toml, as specified by the documentation
c. the example config - example.toml, a whole example config provided also by the website documentation
- bashrc (bash configs)

edits:
-  hyprland.conf
a. edited $statusBar to add instructions on switching between status bars
b. edited $mainMod + F5 to killall all possible status bars while still allowing to reset them
c. added F5 bind to start a status bar (mostly used after killing said status bar process)
- links.sh
a. now integrates a script that can check for link names (indiscriminate, or symlink-specific)
b. temporarily commented out symlink creation for bashrc and zsh
c. added ashell to list of folders to add symlinks

TODO:
- zsh installation and configs

+++ links.sh
@@ -88,28 +88,39 @@ LINK="$HOME/.config/starship.toml"
 
 echo -e "${GREEN}Creating symlink for starship...${NC}"
 
-ln -s "$SRC" "$LINK"
+# ln -s "$SRC" "$LINK"
+
+# Use this if checking for a link name (be it a file, symlink, or directory)
+# if [ ! -e "$LINK" ]; then
+#   ln -s "$SRC" "$LINK"
+# fi
+
+# Use this instead if you want to check specifically for a symlink
+if [ ! -L "$LINK" ]; then
+  ln -s "$SRC" "$LINK"
+fi
+
 
 # ---------
 
-SRC="$SCRIPT_DIR/bash/.bashrc"
-LINK="$HOME/.bashrc"
+# SRC="$SCRIPT_DIR/bash/.bashrc"
+# LINK="$HOME/.bashrc"
 
-echo -e "${GREEN}Creating symlink for bash...${NC}"
+# echo -e "${GREEN}Creating symlink for bash...${NC}"
 
-ln -s "$SRC" "$LINK"
+# ln -s "$SRC" "$LINK"
 
 # optionally:
 # source ~/.bashrc
 
 # --------
 
-SRC="$SCRIPT_DIR/zsh/.zshrc"
-LINK="$HOME/.zsh"
+# SRC="$SCRIPT_DIR/zsh/.zshrc"
+# LINK="$HOME/.zsh"
 
-echo -e "${GREEN}Creating symlink for zsh...${NC}"
+# echo -e "${GREEN}Creating symlink for zsh...${NC}"
 
-ln -s "$SRC" "$LINK"
+# ln -s "$SRC" "$LINK"
 
 # optionally:
 # source ~/.zshrc
@@ -121,6 +132,7 @@ FOLDERS=(
   waybar
   hypr/hyprlock
   alacritty
+  ashell
 )
 
 for f in "${FOLDERS[@]}"; do
